Move quoting methods to query builder
